# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/players": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Retrieves all players",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Player"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Creates a Player",
                "parameters": [
                    {
                        "description": "Player",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Player"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/players/squadnumber/{squadnumber}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Retrieves a Player by its Squad Number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player.SquadNumber",
                        "name": "squadnumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Player"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/players/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Retrieves a Player by its internal UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player.ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Player"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/players/{squadnumber}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Updates (entirely) a Player by its Squad Number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player.SquadNumber",
                        "name": "squadnumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Player",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Player"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "players"
                ],
                "summary": "Deletes a Player by its Squad Number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Player.SquadNumber",
                        "name": "squadnumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Player": {
            "type": "object",
            "properties": {
                "abbrPosition": {
                    "description": "The abbreviated form of the Player's position",
                    "type": "string"
                },
                "dateOfBirth": {
                    "description": "The date of birth of the Player",
                    "type": "string"
                },
                "firstName": {
                    "description": "The first name of the Player",
                    "type": "string"
                },
                "id": {
                    "description": "Internal UUID (server-generated, opaque to clients)",
                    "type": "string"
                },
                "lastName": {
                    "description": "The last name of the Player",
                    "type": "string"
                },
                "league": {
                    "description": "The league where the team plays",
                    "type": "string"
                },
                "middleName": {
                    "description": "The middle name of the Player, if any",
                    "type": "string"
                },
                "position": {
                    "description": "The playing position of the Player",
                    "type": "string"
                },
                "squadNumber": {
                    "description": "User-facing unique identifier; DB-enforced uniqueness",
                    "type": "integer"
                },
                "starting11": {
                    "description": "Indicates whether the Player is in the starting 11",
                    "type": "boolean"
                },
                "team": {
                    "description": "The team to which the Player belongs",
                    "type": "string"
                }
            }
        }
    }
}